refactor: change redux file structure
