+++ src/hooks/useTypedSelector.ts
@@ -1,4 +1,4 @@
 import { TypedUseSelectorHook, useSelector } from 'react-redux';
-import { RootState } from '../store/reducers';
+import { RootState } from '../store/rootReducer';
 
 export const useTypedSelector: TypedUseSelectorHook<RootState> = useSelector;
