+++ src/store/auth/authReducer.ts
@@ -7,7 +7,7 @@ import {
   SetTokenInterface,
   SetUserInterface,
 } from './types';
-import { UserInterface } from '../../../models/UserInterface';
+import { UserInterface } from '../../models/UserInterface';
 
 export interface AuthState {
   isAuth: boolean;
