+++ src/store/index.ts
@@ -1,6 +1,6 @@
 import { applyMiddleware, createStore } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
-import { rootReducer } from './reducers';
+import { rootReducer } from './rootReducer';
 
 export const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk)));
