+++ src/store/lots/lotsReducer.ts
