+++ src/store/lots/types.ts
